# project/doc/openapi.yaml
openapi: 3.0.0
info:
  description: | 
    API Specification
  version: "1.0.0"
  title: Squares
  contact:
    email: ilyalyov@mail.com
    name: Ilya
servers:
  - url: http://localhost:8000/api/v1
    description: Local dev server
...
paths:
...
  /users/{userId}:
    get:
      summary: Get user
      description: This can only be done by authorised user.
      operationId: getUser
      parameters:
        - in: path
          name: userId
          schema:
            type: integer
            format: int64
          required: true
      responses:
        '200':
          description: 'OK'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
      tags:
        - user
    patch:
      summary: Update user
      description: This can only be done by authorized user.
      operationId: updateUser
      parameters:
        - in: path
          name: userId
          schema:
            type: integer
            format: int64
          required: true
      requestBody:
        description: User to update
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User_update'
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
      tags:
        - user
...
components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
        user:
          type: object
          properties:
            username:
              type: string
            firstName:
              type: string
            lastName:
              type: string
            email:
              type: string
            password:
              type: string
              writeOnly: true
        phone:
          type: string
        about:
          type: string
        photo:
          type: string
      required:
        - username
        - email
        - password
...
    User_update:
      type: object
      properties:
        user:
          type: object
          properties:
            username:
              type: string
            firstName:
              type: string
            lastName:
              type: string
            email:
              type: string
            password:
              type: string
              writeOnly: true
        phone:
          type: string
        about:
          type: string
        photo:
          type: string
          format: binary
...